ingester: add collection (un)verify support
This commit introduces collectiong verification logic to the ingester
code, as well as a new migration for the database which adds two new
columns to the `asset` table, namely `collection_verified` and
`collection`. I also used `sea-orm-cli` to regenerate the bindings
and wrappers for the `asset` table, so we can work with the new fields
programmatically.

Signed-off-by: Alexandru Agache <alexandruag@gmail.com>