Updated tools/tests (#99)
* Adds a reindex param

The reindex param is used to indicate the need for reindexing. This fixes an issue where the reindexing procedure would reset the field to processing, removing the stale but valid data that already existed in the column. It also makes it easier to find all the assets that require reindexing by having an indexed field to search for.

* Exported libs for the nft ingester

* Upstreaming tools changes from Triton

Upstreaming the changes made to the tools from Triton's branch.

* update tools to remove need for custom solana

* Exported libs for the nft ingester

* Upstreaming tools changes from Triton

Upstreaming the changes made to the tools from Triton's branch.

* update tools to remove need for custom solana

* Added cargo lock and fixed build

* Rename tests to tools

* Delete migration

* move back'